[add utils to diagram]
