+++ packages/diagram/package.json
@@ -7,5 +7,8 @@
   "repository": "https://github.com/Leooman/LhnUI.git",
   "author": "Leooman <yonghenglion@gmail.com>",
   "license": "MIT",
-  "private": true
+  "private": true,
+  "dependencies": {
+    "@lhn/utils": "^0.0.0"
+  }
 }
