+++ packages/diagram/src/utils/utils.ts
@@ -34,24 +34,3 @@ export function pointInLine(point: Point, from: Point, to: Point): boolean {
 
   return pointInRect(point, points)
 }
-
-function S4(): string {
-  return (((1 + Math.random()) * 0x10000) | 0).toString(16).substring(1)
-}
-
-export function uuid(): string {
-  return (
-    S4() +
-    S4() +
-    '-' +
-    S4() +
-    '-' +
-    S4() +
-    '-' +
-    S4() +
-    '-' +
-    S4() +
-    S4() +
-    S4()
-  )
-}
